machine: add firewall_enabled
This is a bit hacky at the moment because the gosdc API doesn't expose
an attribute telling us if the firewall is enabled for a machine or not.
This means that we can set information about the firewall, but we don't
know for sure whether it's on or off when the state is refreshed.